Create PurchaseListViewController showing purchase list
Separate the purchase list view by creating a content view.
